Display profile pic of sender in notification bubbles
This is a little wrong - the profile pics are fetched synchronously for now,
meaning the UI is blocked while they're being downloaded. Putting in the
callbacks and such to fetch them asynchronously will be tremendously messy. It's
OK for now because the pics download so fast that you don't notice a delay.