Fix memory leak (this is the proper fix for the animation crash)
I don't fully understand why copying the animation (which was a leak) fixed the
crash. The trouble was the window getting over-released, since -close releases
the window by default. It seems very unlikely that the animation would be
releasing the window - it shouldn't be retaining the window in the first place -
so it seems like the animation should have nothing to do with the crash. This
fix is pretty clear, though.

The underlying problem is that the design of Cocoa doesn't anticipate this -
usually, windows are not responsible for closing themselves.